+++ lib/zenlish/wclasses/modal_verb_could.rb
@@ -1,0 +1,9 @@
+require_relative 'auxiliary'
+
+module Zenlish
+  module WClasses
+    # The modal verb `could`
+    class ModalVerbCould < Auxiliary
+    end # class
+  end # module
+end # module